content-moderation/requirements.txt
torch
numpy
pydantic
daily-python
transformers